refactor: extract recording events from App
