+++ src/app/App.tsx
@@ -334,6 +334,8 @@ import { useActiveProfileData } from "./use-active-profile-data";
 
 import { useRecordingStartActions } from "./use-recording-start-actions";
 
+import { useRecordingEvents } from "./use-recording-events";
+
 export function App() {
   const replayOnboarding = shouldReplayOnboarding();
   const activeHermesProfileName = useActiveHermesProfileName();
@@ -3053,111 +3055,13 @@ export function App() {
     drainPendingMeetingStartRef.current();
   }, [appBlocked, bootstrapped]);
 
-  useEffect(() => {
-    let aborted = false;
-    let unlisten: (() => void) | undefined;
-    let listenerRetryTimer: number | undefined;
-    let drainRetryTimer: number | undefined;
-    let drainRunning = false;
-    let drainAgain = false;
-
-    const scheduleDrainRetry = () => {
-      if (aborted || drainRetryTimer !== undefined) return;
-      drainRetryTimer = window.setTimeout(() => {
-        drainRetryTimer = undefined;
-        drainPendingRequest();
-      }, 500);
-    };
-
-    function drainPendingRequest() {
-      if (!meetingStartReadyRef.current || !meetingStartListenerRegisteredRef.current) return;
-      if (drainRunning) {
-        drainAgain = true;
-        return;
-      }
-      drainRunning = true;
-      void (async () => {
-        let shouldRetry = false;
-        try {
-          const request = await pendingMeetingStartRequest();
-          if (
-            aborted ||
-            !meetingStartReadyRef.current ||
-            !meetingStartListenerRegisteredRef.current ||
-            !request
-          ) {
-            return;
-          }
-          if (request.expired) {
-            revealMainWindowForMeetingStartError();
-            setError(MEETING_START_REQUEST_EXPIRED_MESSAGE);
-            const acknowledged = await acknowledgeMeetingStartRequest(request.requestId);
-            shouldRetry = !acknowledged;
-            return;
-          }
-          const terminal = await meetingStartHandlerRef.current(request.requestId, request.noteId);
-          if (
-            aborted ||
-            !meetingStartReadyRef.current ||
-            !meetingStartListenerRegisteredRef.current
-          ) {
-            return;
-          }
-          if (!terminal) {
-            shouldRetry = true;
-            return;
-          }
-          const acknowledged = await acknowledgeMeetingStartRequest(request.requestId);
-          shouldRetry = !acknowledged;
-        } catch {
-          shouldRetry = true;
-        } finally {
-          drainRunning = false;
-          if (!aborted) {
-            if (drainAgain) {
-              drainAgain = false;
-              drainPendingRequest();
-            } else if (shouldRetry) {
-              scheduleDrainRetry();
-            }
-          }
-        }
-      })();
-    }
-    drainPendingMeetingStartRef.current = drainPendingRequest;
-
-    const register = (attempt = 0) => {
-      void listen(MEETING_START_TRANSCRIPTION_EVENT, drainPendingRequest)
-        .then((cleanup) => {
-          if (aborted) {
-            cleanup();
-            return;
-          }
-          unlisten = cleanup;
-          meetingStartListenerRegisteredRef.current = true;
-          drainPendingRequest();
-        })
-        .catch((error) => {
-          if (!aborted) {
-            console.warn("Failed to register the meeting start listener; retrying.", error);
-            const retryDelay =
-              MEETING_START_LISTENER_RETRY_DELAYS_MS[
-                Math.min(attempt, MEETING_START_LISTENER_RETRY_DELAYS_MS.length - 1)
-              ];
-            listenerRetryTimer = window.setTimeout(() => register(attempt + 1), retryDelay);
-          }
-        });
-    };
-
-    register();
-    return () => {
-      aborted = true;
-      meetingStartListenerRegisteredRef.current = false;
-      if (listenerRetryTimer !== undefined) window.clearTimeout(listenerRetryTimer);
-      if (drainRetryTimer !== undefined) window.clearTimeout(drainRetryTimer);
-      unlisten?.();
-    };
-  }, []);
+  useRecordingEvents({
+    drainPendingMeetingStartRef,
+    meetingStartHandlerRef,
+    meetingStartListenerRegisteredRef,
+    meetingStartReadyRef,
+    setError,
+  });
 
   // The handler closes over frequently-changing state, but the Tauri listener
   // must register exactly once: re-subscribing tears the listener down and
